refactor: moved water functionality to feature first folder structure
