+++ lib/features/auth/providers/user/user_provider.dart
@@ -1,0 +1,10 @@
+/// ------ User and Profile Providers ------ ///
+final UserService userService = UserService();
+final UserLocalService userLocalService = UserLocalService();
+final UserRepository userRepository = UserRepository(
+    userService: userService,
+    userLocalService: userLocalService,
+    authService: authService);
+
+final userProvider = AsyncNotifierProvider<UserNotifier, AppUser>(
+        () => UserNotifier(userRepository: userRepository));
