+++ lib/features/auth/screens/register_screen.dart
@@ -8,7 +8,7 @@ import 'package:sickler/models/models.dart';
 import 'package:sickler/features/auth/screens/sign_in_screen.dart';
 import 'package:sickler/screens/global_components/components.dart';
 
-import '../../../providers/providers.dart';
+import '../providers/auth_providers.dart';
 import '../../../screens/global_components/bottom_nav_bar.dart';
 import '../../../screens/profile/profile_basic_info_screen.dart';
 import 'auth_success.dart';
