+++ lib/features/water/models/water_preferences.dart
