+++ lib/features/water/models/water_stats.dart
