+++ lib/features/water/providers/water_log_notifier.dart
@@ -5,8 +5,8 @@ import 'package:fpdart/fpdart.dart';
 import 'package:sickler/models/water/water_log.dart';
 import 'package:sickler/repositories/water/water_repository.dart';
 
-import '../../core/core.dart';
-import '../../features/auth/models/app_user.dart';
+import '../../../core/core.dart';
+import '../../auth/models/app_user.dart';
 
 class WaterLogNotifier extends AsyncNotifier<List<WaterLog>> {
   final WaterRepository _waterRepository;
