+++ lib/features/water/screens/suggested_water_daily_goal_screen.dart
@@ -6,7 +6,7 @@ import 'package:go_router/go_router.dart';
 import 'package:sickler/core/core.dart';
 import 'package:sickler/models/models.dart';
 import 'package:sickler/models/water/water_preferences.dart';
-import 'package:sickler/providers/providers.dart';
+import 'package:sickler/features/auth/providers/auth_providers.dart';
 import 'package:sickler/providers/water/water_providers.dart';
 import 'package:sickler/screens/global_components/components.dart';
