+++ lib/screens/home/home_screen.dart
@@ -4,7 +4,7 @@ import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 import 'package:sickler/features/auth/models/app_user.dart';
 import 'package:sickler/models/water/water_preferences.dart';
-import 'package:sickler/providers/providers.dart';
+import 'package:sickler/features/auth/providers/auth_providers.dart';
 import 'package:sickler/providers/water/water_providers.dart';
 import 'package:sickler/screens/emergency/emergency_screen.dart';
 import 'package:sickler/screens/profile/profile_screen.dart';
